+++ R/CreateArgFunctions.R
@@ -36,6 +36,8 @@
 #'                                         to start ineligible included visits
 #' @param exclusionEvaluationDaysFromEnd   The number of days from the cohort end date of the exclusionEvaluationCohortId
 #'                                         to end ineligible included visits
+#' @param minimumOffsetFromStart           Minimum number of days to offset for the analysis visit from the start of the observation period
+#' @param minimumOffsetFromEnd             Minimum number of days to offset for the analysis visit from the end of the observation period
 #' @param modelBaseSampleSize              The number of non-xSpec subjects to include in the model
 #' @param baseSampleSize                   The maximum number of subjects in the evaluation cohort.
 #' @param lowerAgeLimit                    The lower age for subjects in the model.
@@ -76,7 +78,9 @@ createCreateEvaluationCohortArgs <- function(xSpecCohortId,
                                              exclusionEvaluationCohortId = 0,
                                              exclusionEvaluationDaysFromStart = 0,
                                              exclusionEvaluationDaysFromEnd = 0,
-                                             modelBaseSampleSize = 15000,
+                                             minimumOffsetFromStart = 365,
+                                             minimumOffsetFromEnd = 365,
+                                             modelBaseSampleSize = 25000,
                                              baseSampleSize = 2e+06,
                                              lowerAgeLimit = 0,
                                              upperAgeLimit = 120,