+++ R/CreateEvaluationCohort.R
@@ -29,9 +29,11 @@
 #' @param tempEmulationSchema Some database platforms like Oracle and Impala do not truly support temp tables. To
 #'                            emulate temp tables, provide a schema with write privileges where temp tables
 #'                            can be created.
+#' @param phenotype                        Name of the phenotype for analysis
+#' @param databaseId                       Name of the database in the analysis
 #' @param xSpecCohortId                    The number of the "extremely specific (xSpec)" cohort
 #'                                         definition id in the cohort table (for noisy positives).
-#' @param daysFromxSpec                    Number of days from xSpec condition until analyzed visit
+#' @param daysFromxSpec                    Number of days allowed from xSpec condition until analyzed visit
 #' @param xSensCohortId                    The number of the "extremely sensitive (xSens)" cohort
 #'                                         definition id in the cohort table (for noisy negatives).
 #' @param prevalenceCohortId               The number of the cohort definition id to determine the
@@ -72,6 +74,8 @@
 #'                                         to start ineligible included visits
 #' @param exclusionEvaluationDaysFromEnd   The number of days from the cohort end date of the exclusionEvaluationCohortId
 #'                                         to end ineligible included visits
+#' @param minimumOffsetFromStart           Minimum number of days to offset for the analysis visit from the start of the observation period
+#' @param minimumOffsetFromEnd             Minimum number of days to offset for the analysis visit from the end of the observation period
 #' @param modelBaseSampleSize              The number of non-xSpec subjects to include in the model
 #' @param baseSampleSize                   The maximum number of subjects in the evaluation cohort.
 #' @param lowerAgeLimit                    The lower age for subjects in the model.
@@ -86,6 +90,7 @@
 #' @param falsePositiveNegativeSubjects    Number of subjects to include for evaluating false positives and negatives
 #' @param cdmVersion                       The CDM version of the database.
 #' @param outFolder                        The folder where the output files will be written.
+#' @param exportFolder                     The folder where the csv output files will be written.
 #' @param modelId                          A string used to generate the file names for this model.
 #' @param evaluationCohortId               A string used to generate the file names for this evaluation cohort.
 #' @param excludeModelFromEvaluation       Should subjects used in the model be excluded from the evaluation cohort?
@@ -98,6 +103,8 @@
 createEvaluationCohort <- function(connectionDetails,
                                    oracleTempSchema = NULL,
                                    tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
+                                   phenotype,
+                                   databaseId,
                                    xSpecCohortId,
                                    daysFromxSpec = 0,
                                    xSensCohortId,
@@ -120,7 +127,9 @@ createEvaluationCohort <- function(connectionDetails,
                                    exclusionEvaluationCohortId = 0,
                                    exclusionEvaluationDaysFromStart = 0,
                                    exclusionEvaluationDaysFromEnd = 0,
-                                   modelBaseSampleSize = 15000,
+                                   minimumOffsetFromStart = 365,
+                                   minimumOffsetFromEnd = 365,
+                                   modelBaseSampleSize = 25000,
                                    baseSampleSize = 2e+06,
                                    lowerAgeLimit = 0,
                                    upperAgeLimit = 120,
@@ -134,6 +143,7 @@ createEvaluationCohort <- function(connectionDetails,
                                    falsePositiveNegativeSubjects = 10,
                                    cdmVersion = "5",
                                    outFolder = getwd(),
+                                   exportFolder,
                                    modelId = "main",
                                    evaluationCohortId = "main",
                                    excludeModelFromEvaluation = FALSE,
@@ -173,6 +183,8 @@ createEvaluationCohort <- function(connectionDetails,
   start <- Sys.time()
   .createPhenotypeModel(
     connectionDetails = connectionDetails,
+    phenotype = phenotype,
+    databaseId = databaseId,
     cdmDatabaseSchema = cdmDatabaseSchema,
     cohortDatabaseSchema = cohortDatabaseSchema,
     cohortTable = cohortTable,
@@ -188,6 +200,8 @@ createEvaluationCohort <- function(connectionDetails,
     modelPopulationCohortIdStartDay = modelPopulationCohortIdStartDay,
     modelPopulationCohortIdEndDay = modelPopulationCohortIdEndDay,
     modelBaseSampleSize = modelBaseSampleSize,
+    minimumOffsetFromStart = minimumOffsetFromStart,
+    minimumOffsetFromEnd = minimumOffsetFromEnd,
     lowerAgeLimit = lowerAgeLimit,
     upperAgeLimit = upperAgeLimit,
     visitLength = visitLength,
@@ -200,11 +214,14 @@ createEvaluationCohort <- function(connectionDetails,
     removeSubjectsWithFutureDates = removeSubjectsWithFutureDates,
     cdmVersion = cdmVersion,
     outFolder = outFolder,
+    exportFolder = exportFolder,
     modelId = modelId
   )
 
   .createEvaluationCohort(
     connectionDetails = connectionDetails,
+    phenotype = phenotype,
+    databaseId = databaseId,
     xSpecCohortId = xSpecCohortId,
     xSensCohortId = xSensCohortId,
     prevalenceCohortId = prevalenceCohortId,
@@ -220,6 +237,8 @@ createEvaluationCohort <- function(connectionDetails,
     exclusionEvaluationCohortId = exclusionEvaluationCohortId,
     exclusionEvaluationDaysFromStart = exclusionEvaluationDaysFromStart,
     exclusionEvaluationDaysFromEnd = exclusionEvaluationDaysFromEnd,
+    minimumOffsetFromStart = minimumOffsetFromStart,
+    minimumOffsetFromEnd = minimumOffsetFromEnd,
     baseSampleSize = baseSampleSize,
     lowerAgeLimit = lowerAgeLimit,
     upperAgeLimit = upperAgeLimit,
@@ -233,6 +252,7 @@ createEvaluationCohort <- function(connectionDetails,
     falsePositiveNegativeSubjects = falsePositiveNegativeSubjects,
     cdmVersion = cdmVersion,
     outFolder = outFolder,
+    exportFolder = exportFolder,
     modelId = modelId,
     evaluationCohortId = evaluationCohortId,
     excludeModelFromEvaluation = excludeModelFromEvaluation,